102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,7 +12,7 @@ int main(void)
 {
 	unsigned long int num1 = 1;
 	unsigned long int num2 = 2;
-	unsigned long long int num3;
+	unsigned long int num3;
 	int i;
 
 	printf("%lu, %lu, ", num1, num2);
